Refactor: Moved some functions to a more appropriate file name
