+++ backend/index.js
@@ -9,6 +9,7 @@ const authRoutes = require('./routes/authRoutes');
 const productRoutes = require('./routes/productRoutes')
 const orderRoutes = require('./routes/orderRoutes')
 const checkoutRoutes = require('./routes/checkoutRoutes')
+const userInfoRoutes = require('./routes/userInfoRoutes')
 
 // Create an Express app
 const app = express();
@@ -34,3 +35,4 @@ app.use('/auth', authRoutes)
 app.use('/products', productRoutes)
 app.use('/order', orderRoutes)
 app.use('/checkout', checkoutRoutes)
+app.use('/user-info', userInfoRoutes)
