+++ backend/routes/userInfoRoutes.js
@@ -1,0 +1,7 @@
+const express = require('express')
+
+const router = express.Router()
+
+router.post('/set-delivery-address', setDeliveryInfo)
+
+module.exports = router
